pyproject.toml
[project]
name = "journal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "ruff>=0.15.5",
    "tortoise-orm[asyncpg]>=1.1.6",
]

[dependency-groups]
dev = [
    "aerich[toml]>=0.8.1",
    "asgi-lifespan>=2.1.0",
    "black>=26.3.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-html>=4.2.0",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"

[tool.aerich]
tortoise_orm = "app.db.base.TORTOISE_ORM"
location = "app/db/migrations"
src_folder = "./."
